raspberry_pi/requirements.txt
# Smart Brewery - Raspberry Pi Dependencies

# Azure IoT Hub SDK
azure-iot-device>=2.12.0

# HTTP requests
requests>=2.28.0

# For BME280 sensor (optional)
# adafruit-circuitpython-bme280

# For DS18B20 temperature sensor (optional)
# w1thermsensor

# For HX711 weight sensor (optional)
# hx711

